requirements.txt
requests>=2.31.0
websocket-client>=1.7.0
tqdm>=4.66.0
